requirements.txt
google-genai
beautifulsoup4
requests
pandas
python-dotenv
matplotlib
gradio
plotly>=5.18.0
scikit-learn>=1.3.0
numpy>=1.24.0
